Keep interest rate and inflator as uint160, timestamps as uint96

+++ src/erc20/ERC20Pool.sol
@@ -17,7 +17,7 @@ contract ERC20Pool is ReentrancyGuard, IERC20Pool, FlashloanablePool {
     /*** State Variables ***/
     /***********************/
 
-    uint256 public override collateralScale;
+    uint128 public override collateralScale;
 
     /****************************/
     /*** Initialize Functions ***/
@@ -29,12 +29,12 @@ contract ERC20Pool is ReentrancyGuard, IERC20Pool, FlashloanablePool {
     ) external override {
         if (poolInitializations != 0) revert AlreadyInitialized();
 
-        collateralScale = collateralScale_;
+        collateralScale = uint128(collateralScale_);
 
-        inflatorSnapshot           = 10**18;
-        lastInflatorSnapshotUpdate = block.timestamp;
-        interestRate               = rate_;
-        interestRateUpdate         = block.timestamp;
+        inflatorSnapshot           = uint160(10**18);
+        lastInflatorSnapshotUpdate = uint96(block.timestamp);
+        interestRate               = uint160(rate_);
+        interestRateUpdate         = uint96(block.timestamp);
 
         loans.init();
